MongoDB: Refactor software tests
